requirements.txt
azure-core==1.27.1
azure-storage-blob==12.16.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
Flask==2.3.2
fuzzywuzzy==0.18.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.21.1
MarkupSafe==2.1.3
pycparser==2.21
pydub==0.25.1
pymongo==4.4.0
python-Levenshtein==0.21.1
pytube @ git+https://github.com/alakaza2022/pytube@663f4e6cbdb5e0e01c5631d7dc1a0b0271dab808
rapidfuzz==3.1.1
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
